requirements.txt
pytesseract
Pillow
pyspellchecker
